Rename fix.Rmd to The Early Indicator Experiment.Rmd
This RMarkdown file details the Early Indicator Experiment. A project I created, developed and produced over my summer internship at Statistics Canada. The experiment served the purpose of identifying trends within the data covered in the Macro Growth Analysis Report and analyzing the time-accuracy trade off of the current release date compared against the other days of the tracking period. The experiment resulted in moving the release date back by 2 days as the analysis revealed day 5 to have the optimal numbers of the tracking period.